requirements.txt
beautifulsoup4==4.12.3
gradio==5.7.1
langchain==0.3.9
langchain_community==0.3.8
numpy==2.1.3
pandas==2.2.3
partialjson==0.0.8
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.5.2
openai==0.28.0